// frontend/src/api/model/salesInPeriodDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Luxis API
 * System API for Luxis
 * OpenAPI spec version: 1.0
 */
import type { SaleReturnDto } from './saleReturnDto'

export interface SalesInPeriodDto {
  /** The start date of the period */
  start: string
  /** The end date of the period */
  end: string
  /** The returned sales in the period */
  sales: SaleReturnDto[]
}