// frontend/src/api/model/salesAggregatedByDayDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Luxis API
 * System API for Luxis
 * OpenAPI spec version: 1.0
 */
import type { SaleAggregatedByDayDto } from './saleAggregatedByDayDto'

export interface SalesAggregatedByDayDto {
  start: string
  end: string
  data: SaleAggregatedByDayDto[]
}