// frontend/src/api/model/returnsInPeriodDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Luxis API
 * System API for Luxis
 * OpenAPI spec version: 1.0
 */
import type { ReturnWithResellerDto } from './returnWithResellerDto'

export interface ReturnsInPeriodDto {
  /** The start date of the period */
  start: string
  /** The end date of the period */
  end: string
  /** The products returns in the period */
  returns: ReturnWithResellerDto[]
}